pyproject.toml
[tool.pdm]

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.7.0",
    "isort>=5.12.0",
]
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Marco", email = "marcoantoniorosario.academico@gmail.com"},
]
dependencies = [
    "django>=4.2.6",
    "djangorestframework>=3.14.0",
    "setuptools>=68.2.2",
    "djangorestframework-simplejwt>=5.3.0",
    "python-magic>=0.4.27",
    "Pillow>=10.0.1",
    "drf-spectacular>=0.26.5",
    "python-dotenv>=1.0.0",
    "netifaces>=0.11.0",
    "gunicorn>=21.2.0",
    "whitenoise[brotli]>=6.5.0",
    "psycopg2-binary>=2.9.9",
    "django-cloudinary-storage>=0.3.0",
    "django-filter>=23.3",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.scripts]
pre_dev = "python ./scripts/set_my_ip.py"
dev = "python manage.py runserver 0.0.0.0:19003"
runserver = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
shell = "python manage.py shell"
test = "python manage.py test"
startapp = "python manage.py startapp {args}"
loaddata = "python manage.py loaddata {args}"
dumpdata = "python manage.py dumpdata {args}"


[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"

